feat: DeviceTable 분리
